+++ src/Consumers/Consumer.php
@@ -10,6 +10,7 @@ use Junges\Kafka\Commit\Contracts\CommitterFactory;
 use Junges\Kafka\Commit\DefaultCommitterFactory;
 use Junges\Kafka\Commit\NativeSleeper;
 use Junges\Kafka\Config\Config;
+use Junges\Kafka\Contracts\CanConsumeMessages;
 use Junges\Kafka\Contracts\KafkaConsumerMessage;
 use Junges\Kafka\Contracts\MessageDeserializer;
 use Junges\Kafka\Exceptions\KafkaConsumerException;
@@ -23,7 +24,7 @@ use RdKafka\Message;
 use RdKafka\Producer as KafkaProducer;
 use Throwable;
 
-class Consumer
+class Consumer implements CanConsumeMessages
 {
     private const IGNORABLE_CONSUMER_ERRORS = [
         RD_KAFKA_RESP_ERR__PARTITION_EOF,
@@ -350,6 +351,10 @@ class Consumer
 
         if ($batchConfig->isBatchingEnabled()) {
             $this->handleBatch();
+        }
+
+        if ($this->config->shouldStopAfterLastMessage() && RD_KAFKA_RESP_ERR__PARTITION_EOF === $message->err) {
+            $this->stopConsume();
         }
 
         if (! in_array($message->err, self::IGNORABLE_CONSUMER_ERRORS)) {