+++ src/Facades/Kafka.php
@@ -3,18 +3,19 @@
 namespace Junges\Kafka\Facades;
 
 use Illuminate\Support\Facades\Facade;
-use Junges\Kafka\Message\Message;
+use Junges\Kafka\Contracts\KafkaProducerMessage;
 use Junges\Kafka\Support\Testing\Fakes\KafkaFake;
 
 /**
  * @method static \Junges\Kafka\Contracts\CanProduceMessages publishOn(string $cluster)
  * @method static \Junges\Kafka\Consumers\ConsumerBuilder consumeUsing(string $consumer)
  * @method static \Junges\Kafka\Consumers\ConsumerBuilder createConsumer(string $brokers, array $topics = [], string $groupId = null)
- * @method static void assertPublished(Message $message = null)
- * @method static void assertPublishedTimes(int $times = 1, Message $message = null)
- * @method static void assertPublishedOn(string $topic, Message $message = null, $callback = null)
- * @method static void assertPublishedOnTimes(string $topic, int $times = 1, Message $message = null, $callback = null)
+ * @method static void assertPublished(KafkaProducerMessage $expectedMessage = null, callable $callback = null)
+ * @method static void assertPublishedTimes(int $times = 1, KafkaProducerMessage $expectedMessage = null, callable $callback = null)
+ * @method static void assertPublishedOn(string $topic, KafkaProducerMessage $expectedMessage = null, callable $callback = null)
+ * @method static void assertPublishedOnTimes(string $topic, int $times = 1, KafkaProducerMessage $expectedMessage = null, callable $callback = null)
  * @method static void assertNothingPublished()
+ * @method static void shouldReceiveMessages(\Junges\Kafka\Contracts\KafkaConsumerMessage|\Junges\Kafka\Contracts\KafkaConsumerMessage[] $messages)
  * @mixin \Junges\Kafka\Kafka
  *
  * @see \Junges\Kafka\Kafka